Move `Dataset` to `stateful`
