+++ packages/engine/lib/orchestrator/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 error = { path = "../../lib/error", features = ["spantrace"] }
 nano = { path = "../nano", default-features = false }
+stateful = { path = "../stateful", default-features = false }
 hash_engine_lib = { path = "../..", default-features = false }
 hash_engine = { path = "../../bin/hash_engine", default-features = false, artifact = "bin" }
