+++ packages/engine/lib/stateful/src/dataset/mod.rs
@@ -1,5 +1,7 @@
 // TODO: DOC: Add module level docs for describing the high level concept of datasets, what they are
 //   and why they exist
 
-pub type Dataset = serde_json::Value;
-pub type DatasetMap = serde_json::Map<String, Dataset>;
+mod segment;
+mod shared;
+
+pub use self::{segment::Dataset, shared::SharedDataset};
