+++ packages/engine/src/datastore/batch/mod.rs
@@ -1,13 +1,11 @@
 mod agent;
 pub mod boolean;
-pub mod dataset;
 pub mod iterators;
 pub mod migration;
 
 pub use stateful::{
     agent::AgentBatch,
     context::{AgentIndex, ContextBatch},
+    dataset::Dataset,
     message::MessageBatch,
 };
-
-pub use self::dataset::Dataset;
