+++ packages/engine/src/fetch.rs
@@ -2,9 +2,10 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use futures::StreamExt;
+use stateful::dataset::SharedDataset;
 
 use crate::{
-    proto::{ExperimentRunRepr, ExperimentRunTrait, FetchedDataset, SharedDataset},
+    proto::{ExperimentRunRepr, ExperimentRunTrait, FetchedDataset},
     Error, Result,
 };
