+++ packages/engine/src/proto.rs
@@ -6,7 +6,7 @@ use std::{
 
 use serde::{Deserialize, Serialize};
 use serde_json::Value as SerdeValue;
-use stateful::globals::Globals;
+use stateful::{dataset::SharedDataset, globals::Globals};
 use uuid::Uuid;
 
 use crate::simulation::status::SimStatus;
@@ -144,30 +144,6 @@ impl EngineStatus {
             EngineStatus::UserWarnings(..) => "UserWarnings",
             EngineStatus::PackageError(..) => "PackageError",
         }
-    }
-}
-
-#[derive(Deserialize, Serialize, Clone)]
-pub struct SharedDataset {
-    pub name: Option<String>,
-    pub shortname: String,
-    pub filename: String,
-    pub url: Option<String>,
-    /// Whether the downloadable dataset is a csv
-    pub raw_csv: bool,
-    pub data: Option<String>,
-}
-
-impl Debug for SharedDataset {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("SharedDataset")
-            .field("name", &self.name)
-            .field("shortname", &self.shortname)
-            .field("filename", &self.filename)
-            .field("url", &self.url)
-            .field("raw_csv", &self.raw_csv)
-            .field("data", &CleanOption(&self.data))
-            .finish()
     }
 }
